debian: add missing python-requests build-depends
python-requests is not necessarily needed to build, but svtplay-dl's
test suite will try to load it, without attempting to mock it in any
way. I'll have a look at getting rid of this build dependency, but for
now, add it to build-depends.

Btw, I don't know why this wasn't caught when building it locally in my
cowbuilder :-(. I was able to reproduce now, so I must have missed something.