__init__: Allow sideeffect free access to argparse object
Add a gen_argument_parser() function so that we easily can intercept the
object before it's being used. This allows for things like generating
manpages directly from the argument definitions (kudos to the
argparse-manpage project).

The argparse-manpage tool requires the function to be callable without
arguments, hence the "unknown" default value for version.

The tool can either be integrated in the release process of svtplay-dl
(affecting the set of dependencies for everybody), or integrated as part
of the distribution building process (limiting the dependency to opt-in
usage from distributions --- or opening up for other solutions). This
change will allow for either.

This change is backwards compatible; the interface or behavior of
parser() does not change.